Return EINVAL when alignment is not a power of 2
Consider to match the standard of posix,
added a check of power of 2 to alignment in posix_memalign.
Since, the check of multiple of sizeof(void*) contains the check
of 0.
I just check if alignment is a power of 2 behind without consider the
condition it is 0.
Splited the line to 2 for it is too long.