Avoid system environment path in find_package
If the project root directory lies in user's home directory,
`find_package` will search the project directory instead of the
installed path.

For example, if one's project directory is /home/user1/mimalloc,
`find_package` will first search the prefix `$PATH`, then search
`CMAKE_INSTALL_PREFIX` if no file is found in previous path.

If one clones the project in `${HOME}` and added `${HOME}/bin` to
`${PATH}` (which is common), the error could happen.

To avoid this, we could simply add NO_SYSTEM_ENVIONMENT_PATH to
the end of `find_package`.