Somewhat reorganize functions to ensure control over block initialization (or lack thereof)
If all "public" functions internally call a helper with the desired "initialization"
mode (zero or unitialized) we can ensure that a block is only filled once.

Otherwise, if a "public" function would call another "public" function, it could happen
that a block is first filled with the "uninit memory" marker in debug mode, after which
it's immediately overwritten with all zeroes.