CMakeLists.txt: check /proc/cpuinfo for an SV39 MMU
If the host has an SV39 MMU, the usual aligned hinting strategy will
not work despite the system being 64-bit. On RISC-V linux systems, the
type of MMU can be read from /proc/cpuinfo. If we find one, a constant
is defined that will allow us to skip the aligned hinting when the
time comes.