src/os.c: avoid doomed alloc/free on 32-bit and SV39 systems
On 32-bit or SV39 systems, _mi_os_get_aligned_hint() in src/os.c is
ifdef'd to always return NULL. As a result, the OS prim allocation
routines using that hint will fail to obtain aligned memory. When that
happens, we free() the result, and overallocate.

But this is wasteful: we know that (with high probability) the first
attempt at an aligned allocation will fail. This commit modifies
mi_os_prim_alloc_aligned() to skip the doomed alloc and subsequent
free using the same ifdef condition that is used to disable the hint.